+++ lib/grape/endpoint.rb
@@ -32,6 +32,15 @@ module Grape
         superclass.run_before_each(endpoint) unless self == Endpoint
         before_each.each { |blk| blk.try(:call, endpoint) }
       end
+
+      def block_to_unbound_method(block)
+        return unless block
+
+        define_method :temp_unbound_method, block
+        method = instance_method(:temp_unbound_method)
+        remove_method :temp_unbound_method
+        method
+      end
     end
 
     # Create a new endpoint.
@@ -70,7 +79,7 @@ module Grape
       @status = nil
       @stream = nil
       @body = nil
-      @source = block
+      @source = self.class.block_to_unbound_method(block)
       @before_filter_passed = false
     end
 
@@ -190,10 +199,7 @@ module Grape
       return unless @source
 
       ActiveSupport::Notifications.instrument('endpoint_render.grape', endpoint: self) do
-        instance_exec(&@source)
-      rescue LocalJumpError => e
-        Grape.deprecator.warn 'Using `return` in an endpoint has been deprecated. Use `next` instead.'
-        return e.exit_value
+        @source.bind_call(self)
       end
     end