+++ spec/grape/endpoint_spec.rb
@@ -762,27 +762,11 @@ describe Grape::Endpoint do
     expect(last_response.body).to eq('yo')
   end
 
-  context 'when `return`' do
-    it 'calls deprecator' do
+  context 'when calling return' do
+    it 'does not raise a LocalJumpError' do
       subject.get('/home') do
         return 'Hello'
       end
-
-      expect(Grape.deprecator).to receive(:warn).with('Using `return` in an endpoint has been deprecated. Use `next` instead.')
-
-      get '/home'
-      expect(last_response.status).to eq(200)
-      expect(last_response.body).to eq('Hello')
-    end
-  end
-
-  context 'when `next`' do
-    it 'does not call deprecator' do
-      subject.get('/home') do
-        next 'Hello'
-      end
-
-      expect(Grape.deprecator).not_to receive(:warn)
 
       get '/home'
       expect(last_response.status).to eq(200)